+++ library/src/methods/keyof/keyof.ts
@@ -36,13 +36,13 @@ type TupleOrNever<T> = T extends [string, ...string[]] ? T : never;
  * @returns A enum schema.
  */
 export function keyof<
-  TSchema extends ObjectSchema<any> | ObjectSchemaAsync<any>
+  TSchema extends ObjectSchema<any, any> | ObjectSchemaAsync<any, any>
 >(
   schema: TSchema
-): EnumSchema<TupleOrNever<UnionToTuple<keyof TSchema['object']>>> {
+): EnumSchema<TupleOrNever<UnionToTuple<keyof TSchema['object']['entries']>>> {
   return enumType(
-    Object.keys(schema.object) as TupleOrNever<
-      UnionToTuple<keyof TSchema['object']>
+    Object.keys(schema.object.entries) as TupleOrNever<
+      UnionToTuple<keyof TSchema['object']['entries']>
     >
   );
 }