+++ library/src/methods/omit/types.ts
@@ -4,5 +4,8 @@ import type { ObjectSchema, ObjectSchemaAsync } from '../../schemas/index.ts';
  * Object keys type.
  */
 export type ObjectKeys<
-  TObjectSchema extends ObjectSchema<any> | ObjectSchemaAsync<any>
-> = [keyof TObjectSchema['object'], ...(keyof TObjectSchema['object'])[]];
+  TObjectSchema extends ObjectSchema<any, any> | ObjectSchemaAsync<any, any>
+> = [
+  keyof TObjectSchema['object']['entries'],
+  ...(keyof TObjectSchema['object']['entries'])[]
+];