+++ library/src/schemas/tuple/tupleAsync.ts
@@ -8,15 +8,15 @@ import type {
 import {
   executePipeAsync,
   getIssues,
+  getRestAndDefaultArgs,
   getSchemaIssues,
 } from '../../utils/index.ts';
 import type { TupleInput, TupleOutput, TuplePathItem } from './types.ts';
-import { getTupleArgs } from './utils/index.ts';
 
 /**
  * Tuple shape async type.
  */
-export type TupleShapeAsync = [
+export type TupleItemsAsync = [
   BaseSchema | BaseSchemaAsync,
   ...(BaseSchema[] | BaseSchemaAsync[])
 ];
@@ -25,7 +25,7 @@ export type TupleShapeAsync = [
  * Tuple schema async type.
  */
 export type TupleSchemaAsync<
-  TTupleItems extends TupleShapeAsync,
+  TTupleItems extends TupleItemsAsync,
   TTupleRest extends BaseSchema | BaseSchemaAsync | undefined = undefined,
   TOutput = TupleOutput<TTupleItems, TTupleRest>
 > = BaseSchemaAsync<TupleInput<TTupleItems, TTupleRest>, TOutput> & {
@@ -41,7 +41,7 @@ export type TupleSchemaAsync<
  *
  * @returns An async tuple schema.
  */
-export function tupleAsync<TTupleItems extends TupleShapeAsync>(
+export function tupleAsync<TTupleItems extends TupleItemsAsync>(
   items: TTupleItems,
   pipe?: PipeAsync<TupleOutput<TTupleItems, undefined>>
 ): TupleSchemaAsync<TTupleItems>;
@@ -55,7 +55,7 @@ export function tupleAsync<TTupleItems extends TupleShapeAsync>(
  *
  * @returns An async tuple schema.
  */
-export function tupleAsync<TTupleItems extends TupleShapeAsync>(
+export function tupleAsync<TTupleItems extends TupleItemsAsync>(
   items: TTupleItems,
   error?: ErrorMessage,
   pipe?: PipeAsync<TupleOutput<TTupleItems, undefined>>
@@ -71,7 +71,7 @@ export function tupleAsync<TTupleItems extends TupleShapeAsync>(
  * @returns An async tuple schema.
  */
 export function tupleAsync<
-  TTupleItems extends TupleShapeAsync,
+  TTupleItems extends TupleItemsAsync,
   TTupleRest extends BaseSchema | BaseSchemaAsync | undefined
 >(
   items: TTupleItems,
@@ -90,7 +90,7 @@ export function tupleAsync<
  * @returns An async tuple schema.
  */
 export function tupleAsync<
-  TTupleItems extends TupleShapeAsync,
+  TTupleItems extends TupleItemsAsync,
   TTupleRest extends BaseSchema | BaseSchemaAsync | undefined
 >(
   items: TTupleItems,
@@ -100,8 +100,8 @@ export function tupleAsync<
 ): TupleSchemaAsync<TTupleItems, TTupleRest>;
 
 export function tupleAsync<
-  TTupleItems extends TupleShapeAsync,
-  TTupleRest extends BaseSchema | BaseSchemaAsync | undefined
+  TTupleItems extends TupleItemsAsync,
+  TTupleRest extends BaseSchema | BaseSchemaAsync | undefined = undefined
 >(
   items: TTupleItems,
   arg2?:
@@ -112,7 +112,7 @@ export function tupleAsync<
   arg4?: PipeAsync<TupleOutput<TTupleItems, TTupleRest>>
 ): TupleSchemaAsync<TTupleItems, TTupleRest> {
   // Get rest, error and pipe argument
-  const [rest, error, pipe] = getTupleArgs<
+  const [rest, error, pipe] = getRestAndDefaultArgs<
     TTupleRest,
     PipeAsync<TupleOutput<TTupleItems, TTupleRest>>
   >(arg2, arg3, arg4);
@@ -144,11 +144,7 @@ export function tupleAsync<
      */
     async _parse(input, info) {
       // Check type of input
-      if (
-        !Array.isArray(input) ||
-        (!rest && items.length !== input.length) ||
-        (rest && items.length > input.length)
-      ) {
+      if (!Array.isArray(input) || items.length > input.length) {
         return getSchemaIssues(
           info,
           'type',