+++ library/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Add `getRestAndDefaultArgs` utility function
+- Add new `rest` argument to `object` and `objectAsync` schema
+- Change `ObjectSchema` and `ObjectSchemaAsync` type
+- Change type check in `tuple` and `tupleAsync` to be less strict
+- Rename `ObjectShape` and `ObjectShapeAsync` types to `ObjectEntries` and `ObjectEntriesAsync`
+- Rename `TupleShape` and `TupleShapeAsync` types to `TupleItems` and `TupleItemsAsync`
+- Deprecate `passthrough`, `strict` and `strip` method in favor of `object` schema with `rest` argument
+
 ## v0.19.0 (October 08, 2023)
 
 - Add `notBytes`, `notLength`, `notSize` and `notValue` validation function (pull request #194)